refactor: route group setting
