+++ apps/commerce-client-web/src/app/(default)/(layout)/components/footer.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const Footer: React.FC = () => {
+  return (
+    <footer className="w-full bg-background p-4 text-center">
+      <p>&copy; 2024 Inner circle 1기 5조</p>
+    </footer>
+  );
+};
+
+export default Footer;
