+++ apps/commerce-client-web/src/app/(default)/page.tsx
@@ -1,0 +1,14 @@
+import SliderBanner from '@/app/(default)/(home)/components/slidebanner';
+
+export const mockImages = [
+  { src: '/mockbannerimage/book1.png', alt: 'Banner 1' },
+  { src: '/mockbannerimage/book2.png', alt: 'Banner 2' },
+];
+
+export default function Home() {
+  return (
+    <>
+      <SliderBanner images={mockImages} />
+    </>
+  );
+}
