--- apps/commerce-client-web/src/app/layout/components/footer.tsx
@@ -1,11 +1,0 @@
-import React from 'react';
-
-const Footer: React.FC = () => {
-    return (
-        <footer className="w-full p-4 bg-background text-center">
-            <p>&copy; 2024 Inner circle 1기 5조</p>
-        </footer>
-    );
-};
-
-export default Footer;
