--- apps/commerce-client-web/src/app/layout/components/search.tsx
@@ -1,30 +1,0 @@
-import React from 'react';
-
-const Search: React.FC = () => {
-    return (
-        <div className="relative w-full max-w-md mx-auto">
-            <input
-                className="flex h-9 w-full rounded-md border border-input bg-transparent px-3 py-2 text-sm shadow-sm placeholder:text-muted-foreground focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring disabled:cursor-not-allowed disabled:opacity-50"
-                placeholder="Search..."
-                type="search"
-            />
-            <button className="absolute right-3 top-2.5">
-                <svg
-                    xmlns="http://www.w3.org/2000/svg"
-                    viewBox="0 0 24 24"
-                    fill="none"
-                    stroke="currentColor"
-                    strokeWidth="2"
-                    strokeLinecap="round"
-                    strokeLinejoin="round"
-                    className="feather feather-search w-4 h-4 text-gray-400"
-                >
-                    <circle cx="11" cy="11" r="8"></circle>
-                    <line x1="21" y1="21" x2="16.65" y2="16.65"></line>
-                </svg>
-            </button>
-        </div>
-    );
-};
-
-export default Search;
